+++ app/src/main/java/io/github/sibmaks/spring/jfr/report/connections/dto/ConnectionEventDto.java
@@ -20,7 +20,7 @@ public class ConnectionEventDto {
     private ConnectionException exception;
     private long startedAt;
     private long finishedAt;
-    private long threadName;
+    private int threadName;
     private Integer transactionIsolation;
 
 }
